COMP: Avoid setting CMAKE_BACKWARDS_COMPATIBILITY
CMake creates this cache entry whenever it encounters code not known to
require CMake 2.6 or higher in case compatibility with CMake 2.4 or
lower is needed.  Fix all cmake_minimum_required(VERSION) calls in the
ITK tree to be at lest CMake 2.6 even in third party code.  Since the
top of the tree requires CMake 2.8 this will not change the requirement.