Fixup openjpeg situation
By default gdcm will be using the robust 1.x branch of openjpeg which seems to be the most well maintained library. By default a convenience OpenJPEG v1.4 (r697) will be used.
However to be able to continue working on the gsoc project Whole Slide Imaging, we are required to use an OpenJPEG v2 release. Therefore both release both 1.4 and 2.x are now shipped in.
To activate OpenJPEG v2, a user will have to choose at cmake configuration time the cmake variable GDCM_USE_OPENJPEG_V2 explicitely to ON