Add a new API to gdcm::IPPSorter
This new API will allow advanced user to specify the tolerance not only to the Z spacing,
but also on the Direction Cosines (IOP) itself. This is required in some case
where the IOP is very slightly changing in between each slices and string comparison
is inadapted in those cases.