Support file with a single 0,0 attribute at the very begining of the DataSet
This commit reverts portion of commit 2584eaff. It uses carefully the ReadPreValue
to only read the minimum amount of information.
In a very special case, where a DICOM file has a Preamble but not File Meta Information
we are required to read the first DataElement. In this case we only read the Tag and the Value Length
to determine whether or not this is a DICOM.

The speed is important here, see full thread at:
[Gdcm2] Large performance penalty caused by supporting file with 0, 0 attribute?
https://sourceforge.net/mailarchive/message.php?msg_id=26919802