Adding a new class: ImageRegionReader
This new class allow for a flexible mecanisme where the user can specify any kind of Region.
The implementation is then responsible for loading the minimum information from the image into the user specified buffer.
This allow for reading of very large image in chunks.

ImageRegionReader should be used instead of StreamImageReader which is an incomplete implementation. StreamImageReader will focus on the handling of multiresolution image such as Whole Slide Imaging (and JPEG 2000).