Modified gdcmanon and crypto factory structure
Slightly modified the crypto factory class structure
in order to return pointers and not references. Also
made some modifications to gdcmanon to start intoducing
password based encryption and modified the usage of
getopt_long in the parameters parsing loop in order to
avoid repetition of some code (that is commented for now).