Fix an issue with yet-another broken JPEG/DICOM file
This file was defined with:
(7fe0,0010) OB (PixelSequence #=3)                      # u/l, 1 PixelData
  (fffe,e000) pi (no value available)                     #   0, 1 Item
  (fffe,e000) pi ff\d8\ff\e0\00\0f\4c\4a\49\46\00\01\00\00\00\00\00\00\00\ff\fe\00... # 139656, 1 Item
  (fffe,e000) pi (not loaded)                             # 139656, 1 Item
(fffe,e0dd) na (SequenceDelimitationItem)               #   0, 0 SequenceDelimitationItem

GDCM/JPEG assumed the instance was defined as either a cunked-jpeg and/or a broken multiframes JPEG DICOM/file.
In this case we can assert the reading+decompression went fine and skip the broken (empty) Item.