Provide a fix for BUG: 279
gdcm::Anonymizer class would maintain a mapping of real uids to generated uids as a static map initialized inside a member function.
What this means is that during the whole process execution it would not be possible to ever empty this map.
While very dangerous it seems it was too drastic from an implementation point of view. Instead move the static data to class member
and provide a static function to clear the mapping (with a warning message in documentation).

Thanks to Chris Lampe for report and demo