Fix a crash in gdcm::Sorter
there seems to be a bug in the gdcmSorter.cxx file in Sorter::StableSort .
When a list of filenames (filenames)  is parsed it tries to read all the files and if it fails
it assigns a NULL to the filelist. (lines 80-95)
After this the filelist is sorted using a sort functor, which tries to access the GetDataSet() function
of the objects stored in filelist (line 84)

class SortFunctor
...
return (SortFunc)(file1->GetDataSet(), file2->GetDataSet()); )

which causes a segmentation fault for NULL entries.

Possible patch would be to return false if a file could not be read,
So this event can be handled by the calling function.