Fix an incorrect behavior in StringFilter::FromString
As noticed by the ITK dev team the sf::FromString would not work for multi valued binary attribute.
The for loop was not properly discarding the backslash separator as required.
As part of the initial patch proposed by ITK team another switch case was removed since it would generated dead code, see:
http://review.source.kitware.com/#/c/17037/
for more informations.