ENH: Use new CMake Windows export symbol dll feature.
This allows for all libraries to be built shared on Windows.
Instead of just a few of them. Although this approach does
not require dllimport/dllexport for every symbol it does
require it for static data members. This patch also enables
the export of symbols from the third party libraries that
had support already but was not being used in the ITK build.
To use this feature CMake 3.4 must be used or a master
branch build of CMake as of 7/23/2015. In a few places the
existing import/export code did not work with the new feature,
this redefines those macros to be empty when this feature is on.
This is done with a <library>_EXPORT_INHERIT_TEMPLATE macro.
If you have a class that inherits from a templated class or has
static members that are templates, it causes duplicate symbols
with the new feature and dll builds. In these cases a new .h
file was created that uses the CMake generated EXPORT macro
or sets the macro to empty for use with the dll export feature.
Change-Id: I8c042248a8bbcde2ca1bb9d9b159b658d0f2c5a7