gdcm::FileMetaInformation Read/Write
I came across what appears to be a discrepancy between Write() and
Read() of gdcm::FileMetaInformation:

 Source/DataStructureAndEncodingDefinition/gdcmFileMetaInformation.cxx

It looks like Write() is writing the DICOM preamble, but Read() does not
read it causing an exception when calling Read() on the stream that is
written by Write().

If I skip the first 132 bytes before calling Read(), it seems to work.