ENH: Use CMAKE_INTERPROCEDURAL_OPTIMIZATION:BOOL=ON
The use of CMAKE_INTERPROCEDURAL_OPTIMIZATION:BOOL=ON allows the linker
and the compiler to communicate more closely.

This will cause compilation warnings if any symbol violates
the one-definition-rule (i.e. a symbol is defined in more
than one translation unit with different definitions).

NOTE:  LTO is not supported with clang on linux with the ldgold linker