Change VRType enum to an explicit underlying type
For the long thread see the original pull request at:
https://github.com/malaterre/GDCM/pull/86

GCC seems to properly implement what is mandates in the standard:

[...]
For an enumeration whose underlying type is not fixed, the underlying
type is an integral type that can represent all the enumerator values
defined in the enumeration. If no integral type can represent all the
enumerator values, the enumeration is ill-formed. It is
implementation-defined which integral type is used as the underlying
type except that the underlying type shall not be larger than int unless
the value of an enumerator cannot fit in an int or unsigned int. If the
enumerator-list is empty, the underlying type is as if the enumeration
had a single enumerator with value 0.
[...]
Ref: http://eel.is/c++draft/dcl.enum#7

However some compilers seems to be missing this requirement. Using an
explicit `long long` type here solve the symptoms for those.