BUG: Fix GDCM crash when reading DICOM image
GDCM crashed when a DICOM image that had 32 bits allocated.
The problem was that DoOverlayCleanup returned failure but that return value was ignored and later an empty buffer read was attempted.

This fix avoids the crash by not ignoring the failed overlay cleanup.
Applications can detect the error and switch to DCMTK IO to read such files (DCMTK can correctly read files with 32 bits allocated).