Check for Secondary Capture spacing following DICOM Part 3 Sect A.8.1.3
Following the Secondary Capture Image IOD Module Note:

> If Image Position (Patient) (0020,0032) and Image Orientation (Patient) (0020,0037) (from the Image Plane Module) are present, then the values of Pixel Spacing (0028,0030) (from the Image Plane Module and the Basic Pixel Spacing Calibration Macro included from the SC Image Module) are intended to be used for 3D spatial computations, rather than any values of Nominal Scanned Pixel Spacing (0018,2010) (from the SC Image Module), which may also be present.

However, we make sure to throw a warning in `ImageHelper::GetSpacingTagFromMediaStorage` when the tag for a Secondary Capture MediaStorage is requested and SecondaryCaptureImagePlaneModule is enabled because returning a single tag is insufficient for this requirement. This function should not be called on this type of media storage.

Co-authored-by: Mihail Isakov <mihail.isakov@gmail.com>
Co-authored-by: Mathieu Malaterre <mathieu.malaterre@gmail.com>
Co-authored-by: Steve Pieper <pieper@isomics.com>