Return None instead of empty buffer on jpeg error.
Empty buffer when there is an issue getting the buffer.

I personally think this is a better API into python because
currently when an empty buffer is returned with all 0's you can't
tell if thats becuase of an error in getting the buffer or if the
pixel data is actually all 0, such as on the periphery of a CT
scan.